Add a must_include_parent option to Trail and Index.
It is useful when your logical paths should need to include the last directory name that was appended to trail paths. E.g.

    t = Trail.new '.', { :must_include_parent => true }
    t.append_paths '/opt/some/project1', '/opt/some/where/else/project2'

    t.find 'project1/foobar.txt'        # Instead of 'foobar.txt'
    t.find 'project2/bla/another.txt'   # Instead of 'bla/another.txt'

Doing it this way is cleaner than adding '/some/' and '/some/where/else/' as the Trail paths, since you'd have to double check that you were not accidentally finding files that were in peer directories of project1 and project2.